requirements.txt.txt
fastapi==0.134.0
uvicorn==0.41.0
tensorflow==2.20.0
keras==3.12.1
numpy==2.2.6
pillow==12.1.1
python-multipart==0.0.22
requests==2.32.5
streamlit==1.54.0
